Separated algorithms in groups
